requirements.txt
accelerate==1.10.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-generator==1.10
asyncio-extras==1.3.2
attrs==25.4.0
audioread==3.0.1
av==15.1.0
beautifulsoup4==4.14.2
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.12
contourpy==1.3.3
ctranslate2==4.6.0
cycler==0.12.1
decorator==5.2.1
diskcache==5.6.3
distro==1.9.0
faster-whisper==1.2.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-speech==2.33.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llama_cpp_python==0.3.16
llvmlite==0.45.1
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.7.0
networkx==3.3
noisereduce==3.0.3
notion-client==2.5.0
numba==0.62.1
numpy==2.3.3
onnxruntime==1.23.1
openai==2.3.0
outcome==1.3.0.post0
packaging==25.0
pathlib==1.0.1
pillow==11.3.0
platformdirs==4.4.0
playwright==1.55.0
pluggy==1.6.0
pooch==1.8.2
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
pydub==0.25.1
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyparsing==3.2.5
pyperclip==1.11.0
pypiwin32==223
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyttsx3==2.99
pytweening==1.2.0
pywin32==311
PyYAML==6.0.3
qwen-vl-utils==0.0.14
regex==2025.9.18
requests==2.32.5
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
selenium==4.36.0
sentencepiece==0.2.1
setuptools==70.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
transformers==4.57.1
trio==0.31.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==11.0.3
wsproto==1.2.0
yarl==1.22.0
